+++ src/com/landawn/abacus/util/function/BinaryOperator.java
@@ -14,11 +14,6 @@
 
 package com.landawn.abacus.util.function;
 
-import java.util.Collection;
-import java.util.Comparator;
-import java.util.Map;
-
-import com.landawn.abacus.util.Fn;
 import com.landawn.abacus.util.Try;
 
 /**
@@ -28,24 +23,4 @@ import com.landawn.abacus.util.Try;
  * @author Haiyang Li
  */
 public interface BinaryOperator<T> extends BiFunction<T, T, T>, java.util.function.BinaryOperator<T>, Try.BinaryOperator<T, RuntimeException> {
-
-    static <T, C extends Collection<T>> BinaryOperator<C> ofAddAll() {
-        return Fn.BinaryOperators.ofAddAll();
-    }
-
-    static <T, C extends Collection<T>> BinaryOperator<C> ofRemoveAll() {
-        return Fn.BinaryOperators.ofRemoveAll();
-    }
-
-    static <K, V, M extends Map<K, V>> BinaryOperator<M> ofPutAll() {
-        return Fn.BinaryOperators.ofPutAll();
-    }
-
-    static <T> BinaryOperator<T> minBy(Comparator<? super T> comparator) {
-        return Fn.BinaryOperators.minBy(comparator);
-    }
-
-    static <T> BinaryOperator<T> maxBy(Comparator<? super T> comparator) {
-        return Fn.BinaryOperators.maxBy(comparator);
-    }
 }