+++ src/com/landawn/abacus/util/function/TriFunction.java
@@ -33,29 +33,4 @@ public interface TriFunction<A, B, C, R> extends Try.TriFunction<A, B, C, R, Run
 
         return (a, b, c) -> after.apply(apply(a, b, c));
     }
-
-    /**
-     * Returns the specified instance
-     * 
-     * @param func
-     * @return
-     */
-    public static <A, B, C, R> TriFunction<A, B, C, R> of(final TriFunction<A, B, C, R> func) {
-        N.checkArgNotNull(func);
-
-        return func;
-    }
-
-    public static <A, B, C> TriFunction<A, B, C, Void> create(final TriConsumer<A, B, C> triConsumer) {
-        N.checkArgNotNull(triConsumer);
-
-        return new TriFunction<A, B, C, Void>() {
-            @Override
-            public Void apply(A a, B b, C c) {
-                triConsumer.accept(a, b, c);
-
-                return null;
-            }
-        };
-    }
 }